feat: create default timeline organism
