+++ app/src/pages/Feedback/index.jsx
@@ -1,14 +1,12 @@
 import React, { useCallback, useContext, useState } from 'react';
 import { View } from 'react-native';
-import Timeline from 'react-native-timeline-flatlist';
-import colors from '../../../colors';
 import { FeedbackContext } from '../../store/contexts/feedbackContext';
 import { UserContext } from '../../store/contexts/userContext';
 import { useFocusEffect } from '@react-navigation/native';
 import formatDate from '../../utils/formatDate';
 import { NotFound } from '../../components/organisms/NotFound';
-import { FeedbackItem } from './FeedbackItem';
 import { LoadingContext } from '../../store/contexts/loadingContext';
+import { DefaultTimeline } from '../../components/organisms/DefaultTimeline';
 
 export const FeedbackScreen = () => {
     const { getFeedbackByReceiverId } = useContext(FeedbackContext);
@@ -16,6 +14,12 @@ export const FeedbackScreen = () => {
     const { isLoading } = useContext(LoadingContext);
     const [feedbacks, setFeedbacks] = useState([]);
     const hasFeedbacks = feedbacks.length > 0;
+    const mappedFeedbacks = feedbacks.map((feedback) => ({
+        title: feedback.sender.name,
+        description: feedback.message,
+        time: formatDate(feedback.creationDate, '-').slice(0, 5),
+        icon: feedback.sender.photo,
+    }));
 
     const getData = async () => {
         const data = await getFeedbackByReceiverId(user._id);
@@ -28,34 +32,10 @@ export const FeedbackScreen = () => {
         }, []),
     );
 
-    const mappedFeedbacks = feedbacks.map((feedback) => ({
-        title: feedback.sender.name,
-        description: feedback.message,
-        time: formatDate(feedback.creationDate, '-').slice(0, 5),
-        icon: feedback.sender.photo,
-    }));
-
-    const renderContent = (data, sectionID) => (
-        <FeedbackItem data={data} key={sectionID} />
-    );
-
     return (
         <View className="flex-1  h-full p-4">
             {hasFeedbacks && (
-                <Timeline
-                    data={mappedFeedbacks}
-                    lineWidth={4}
-                    lineColor={colors.primary[300]}
-                    circleColor={colors.primary.DEFAULT}
-                    timeStyle={{
-                        backgroundColor: colors.primary.DEFAULT,
-                        padding: 4,
-                        borderRadius: 12,
-                        color: colors.light,
-                        fontFamily: 'montserrat-semibold',
-                    }}
-                    renderDetail={renderContent}
-                />
+                <DefaultTimeline data={mappedFeedbacks} hasImage/>
             )}
             {!hasFeedbacks && !isLoading && (
                 <NotFound
